requirements.txt
pandas
numpy
scikit-learn
xgboost
pycountry
fastapi
uvicorn
python-dotenv
python-multipart
